requirements.txt
Flask-Caching==1.3.3
Flask-GraphQL==1.4.1
gql==0.1.0
graphene==2.0.1
graphene-sqlalchemy==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
redis==2.10.6
oss==2.4.0